// ui/src/api/types.gen.ts
export type DtoTaskStatusResponse = {
    category?: string;
    color?: string;
    default?: boolean;
    id?: string;
    name?: string;
    position?: number;
    project_id?: string;
};

export type DtoTaskTemplateResponse = {
    assignee_ids?: Array<string>;
    created_at?: string;
    created_by?: string;
    description?: string;
    estimate?: number;
    id?: string;
    name?: string;
    next_run_at?: string;
    priority?: string;
    project_id?: string;
    recurrence_days?: string;
    recurrence_pattern?: string;
    status_id?: string;
    updated_at?: string;
};